refactor(characters): adjust screen characters list
